requirements.txt
streamlit>=1.32.0
requests>=2.31.0
langchain-ollama>=0.1.0
langchain-core>=0.1.0
markdown>=3.5.0
PyPDF2>=3.0.0
python-docx>=1.0.0
